Add loader for insurance_policy on invoice
