+++ server/graphql/core/src/loader/loader_registry.rs
@@ -487,6 +487,12 @@ pub async fn get_loaders(
         },
         async_std::task::spawn,
     ));
+    loaders.insert(DataLoader::new(
+        NameInsuranceJoinLoader {
+            connection_manager: connection_manager.clone(),
+        },
+        async_std::task::spawn,
+    ));
 
     loaders
 }
