+++ server/graphql/core/src/loader/mod.rs
@@ -27,6 +27,7 @@ mod loader_registry;
 mod location;
 mod master_list;
 mod name;
+mod name_insurance_join;
 mod name_row;
 mod packaging_variant;
 mod patient;
@@ -86,6 +87,7 @@ pub use loader_registry::{get_loaders, LoaderMap, LoaderRegistry};
 pub use location::LocationByIdLoader;
 pub use master_list::*;
 pub use name::*;
+pub use name_insurance_join::*;
 pub use name_row::*;
 pub use packaging_variant::*;
 pub use patient::*;
