+++ server/graphql/general/src/queries/insurances.rs
@@ -1,18 +1,14 @@
-use async_graphql::{
-    dataloader::DataLoader, Context, Enum, InputObject, Object, Result, SimpleObject, Union,
-};
-use chrono::NaiveDate;
+use async_graphql::{Context, Enum, InputObject, Result, SimpleObject, Union};
+
 use graphql_core::{
-    loader::InsuranceProviderByIdLoader,
     standard_graphql_error::{validate_auth, StandardGraphqlError},
     ContextExt,
 };
+use graphql_types::types::InsuranceNode;
 use repository::name_insurance_join_row::{
     NameInsuranceJoinRow, NameInsuranceJoinSort, NameInsuranceJoinSortField,
 };
 use service::auth::{Resource, ResourceAccessRequest};
-
-use crate::types::{InsurancePolicyNodeType, InsuranceProviderNode};
 
 #[derive(Enum, Copy, Clone, PartialEq, Eq)]
 #[graphql(rename_items = "camelCase")]
@@ -28,61 +24,6 @@ pub struct InsuranceSortInput {
     /// Sort query result is sorted descending or ascending (if not provided the default is
     /// ascending)
     desc: Option<bool>,
-}
-
-#[derive(PartialEq, Debug)]
-pub struct InsuranceNode {
-    insurance: NameInsuranceJoinRow,
-}
-
-#[Object]
-impl InsuranceNode {
-    pub async fn id(&self) -> &str {
-        &self.row().id
-    }
-
-    pub async fn insurance_provider_id(&self) -> &str {
-        &self.row().insurance_provider_id
-    }
-
-    pub async fn policy_number_person(&self) -> &Option<String> {
-        &self.row().policy_number_person
-    }
-
-    pub async fn policy_number_family(&self) -> &Option<String> {
-        &self.row().policy_number_family
-    }
-
-    pub async fn policy_number(&self) -> &String {
-        &self.row().policy_number
-    }
-
-    pub async fn policy_type(&self) -> InsurancePolicyNodeType {
-        InsurancePolicyNodeType::from_domain(&self.row().policy_type)
-    }
-
-    pub async fn discount_percentage(&self) -> &f64 {
-        &self.row().discount_percentage
-    }
-
-    pub async fn expiry_date(&self) -> &NaiveDate {
-        &self.row().expiry_date
-    }
-
-    pub async fn is_active(&self) -> &bool {
-        &self.row().is_active
-    }
-
-    pub async fn insurance_providers(
-        &self,
-        ctx: &Context<'_>,
-    ) -> Result<Option<InsuranceProviderNode>> {
-        let insurance_provider_id = &self.row().insurance_provider_id;
-
-        let loader = ctx.get_loader::<DataLoader<InsuranceProviderByIdLoader>>();
-        let result = loader.load_one(insurance_provider_id.clone()).await?;
-        Ok(result.map(InsuranceProviderNode::from_domain))
-    }
 }
 
 #[derive(SimpleObject)]
@@ -162,16 +103,6 @@ impl InsuranceConnector {
                 .map(InsuranceNode::from_domain)
                 .collect(),
         }
-    }
-}
-
-impl InsuranceNode {
-    pub fn from_domain(insurance: NameInsuranceJoinRow) -> InsuranceNode {
-        InsuranceNode { insurance }
-    }
-
-    pub fn row(&self) -> &NameInsuranceJoinRow {
-        &self.insurance
     }
 }
